SuperheroSightings/src/main/java/com/sg/superherosightings/model/SuperheroOrganizations.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package com.sg.superherosightings.model;

import java.util.Objects;

/**
 *
 * @author jswan
 */
public class SuperheroOrganizations {

    private Superhero superhero;
    private Organization organization;

    public SuperheroOrganizations() {
        this.superhero = new Superhero();
        this.organization = new Organization();
    }

    public Superhero getSuperhero() {
        return superhero;
    }

    public void setSuperhero(Superhero superhero) {
        this.superhero = superhero;
    }

    public Organization getOrganization() {
        return organization;
    }

    public void setOrganziation(Organization organziation) {
        this.organization = organziation;
    }

    @Override
    public int hashCode() {
        int hash = 3;
        return hash;
    }

    @Override
    public boolean equals(Object obj) {
        if (this == obj) {
            return true;
        }
        if (obj == null) {
            return false;
        }
        if (getClass() != obj.getClass()) {
            return false;
        }
        final SuperheroOrganizations other = (SuperheroOrganizations) obj;
        if (!Objects.equals(this.superhero, other.superhero)) {
            return false;
        }
        if (!Objects.equals(this.organization, other.organization)) {
            return false;
        }
        return true;
    }

}
